vesting: rm support for SDK vesting account types (#666)
* vesting: rm support for SDK vesting account types

* changelog

* comment
